[sec/actions] move constants to a separate file to avoid circular deps (#99439)
* [sec/actions] move constants to a separate file to avoid circular deps

* split out setupSavedObjects import

* remove unused import
