+++ x-pack/plugins/actions/server/authorization/actions_authorization.test.ts
@@ -10,7 +10,10 @@ import { securityMock } from '../../../../plugins/security/server/mocks';
 import { ActionsAuthorization } from './actions_authorization';
 import { actionsAuthorizationAuditLoggerMock } from './audit_logger.mock';
 import { ActionsAuthorizationAuditLogger, AuthorizationResult } from './audit_logger';
-import { ACTION_SAVED_OBJECT_TYPE, ACTION_TASK_PARAMS_SAVED_OBJECT_TYPE } from '../saved_objects';
+import {
+  ACTION_SAVED_OBJECT_TYPE,
+  ACTION_TASK_PARAMS_SAVED_OBJECT_TYPE,
+} from '../constants/saved_objects';
 import { AuthenticatedUser } from '../../../security/server';
 import { AuthorizationMode } from './get_authorization_mode_by_source';
