+++ x-pack/plugins/actions/server/constants/saved_objects.ts
@@ -1,0 +1,10 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+export const ACTION_SAVED_OBJECT_TYPE = 'action';
+export const ALERT_SAVED_OBJECT_TYPE = 'alert';
+export const ACTION_TASK_PARAMS_SAVED_OBJECT_TYPE = 'action_task_params';
