+++ x-pack/plugins/actions/server/feature.ts
@@ -6,7 +6,10 @@
  */
 
 import { i18n } from '@kbn/i18n';
-import { ACTION_SAVED_OBJECT_TYPE, ACTION_TASK_PARAMS_SAVED_OBJECT_TYPE } from './saved_objects';
+import {
+  ACTION_SAVED_OBJECT_TYPE,
+  ACTION_TASK_PARAMS_SAVED_OBJECT_TYPE,
+} from './constants/saved_objects';
 import { DEFAULT_APP_CATEGORIES } from '../../../../src/core/server';
 
 export const ACTIONS_FEATURE = {
