+++ x-pack/plugins/cases/server/client/configure/get_mappings.ts
@@ -8,8 +8,7 @@
 import { SavedObjectsClientContract, Logger } from 'src/core/server';
 import { ActionsClient } from '../../../../actions/server';
 import { ConnectorMappingsAttributes, ConnectorTypes } from '../../../common';
-// eslint-disable-next-line @kbn/eslint/no-restricted-paths
-import { ACTION_SAVED_OBJECT_TYPE } from '../../../../actions/server/saved_objects';
+import { ACTION_SAVED_OBJECT_TYPE } from '../../../../actions/server';
 import { ConnectorMappingsServiceSetup } from '../../services';
 import { CasesClientHandler } from '..';
 import { createCaseError } from '../../common/error';
